refactor(desktop): speak-only is a type
CLAUDE.md promises that a call Luke opens for himself "offers no
microphone track, carries no tools". That promise was kept by
convention: one 2,200-line session class owned a capture device and the
`ask_brain` tool, and the announcer passed `connect({ microphone: false })`,
after which five run-time guards on `#withMicrophone` stood between a
briefing and the device.

The promise is now the type. `RealtimeCall` holds the transport every
call shares; `SpeakOnlyCall` is configured with the ordinary session
document overlaid by `SPEAK_ONLY_SESSION_CONFIG` — `tools: []`,
`tool_choice: "none"` — refuses any tool call that could still reach it,
and has no microphone member to guard; `ConversationCall` is the
subclass that adds the device and the one tool. The mouth's session type
is a `Pick` of `SpeakOnlyCall`, so the announcer's path cannot name a
member the developer's call alone has, and the voice hook holds the two
kinds as two objects with one standing at a time: the developer's press
or ask stands Luke's own call down, as `connect`'s takeover used to.
