--- apps/desktop/src/renderer/voice/realtime-session.ts
@@ -1,2180 +1,0 @@
-import {
-  BRAIN_ASK_PENDING_STATUS,
-  type BrainAppActRequest,
-  type BrainAskResult,
-} from "@sidecar/brain/requests-wire";
-import { TRACE_DIRECTION, type TraceDirection } from "@sidecar/devtrace/vocabulary";
-import type { RealtimeConnection } from "@sidecar/hosted";
-import {
-  ARRIVAL_SPEECH_KIND,
-  ASK_BRAIN_TOOL,
-  arrivalSpeechEvents,
-  BRIEFING_SPEECH_KIND,
-  briefingSpeechEvents,
-  CALENDAR_ONBOARDING_SPEECH_KIND,
-  calendarOnboardingSpeechEvents,
-  cancelResponseEvents,
-  clearInputAudioEvents,
-  clearOutputAudioEvents,
-  decodeRealtimePayload,
-  functionCallFollowUpEvents,
-  type IntroductionLine,
-  inputAudioAppendEvents,
-  inputAudioFormatUpdateEvents,
-  introductionSpeechEvents,
-  outputSpeedUpdateEvents,
-  type ParsedRealtimeFunctionCall,
-  PressAudioBuffer,
-  type ProactiveSpeechTurn,
-  parseRealtimeServerEvent,
-  pushToTalkCommitEvents,
-  REALTIME_CLIENT_EVENT,
-  REALTIME_SERVER_EVENT,
-  REALTIME_STATUS,
-  type RealtimeStatus,
-  realtimeSessionConfig,
-  truncateResponseEvents,
-} from "@sidecar/realtime";
-import { maximumTypedAskLength } from "@sidecar/session";
-import {
-  ACT_RESULT_STATUS,
-  isRecord,
-  positiveInteger,
-  text,
-  type UnparsedWireValue,
-  type WireRecord,
-} from "@sidecar/wire";
-import { voiceExchangeActive } from "#shared/messages/voice-view";
-import {
-  type BuiltRealtimeSessionConfig,
-  createAgentsRealtimeTransport,
-  type SdkRealtimeTransport,
-  type SdkToolCallDetails,
-  type SdkTransportFactory,
-} from "./agents-realtime-transport";
-import { MICROPHONE_PROCESSING } from "./microphone-choice";
-import {
-  createPressCaptureSource,
-  type PressCaptureFactory,
-  type PressCaptureSource,
-} from "./press-audio-capture";
-
-/** Bounds the SDK's WebRTC handshake and initial session acknowledgement. */
-const CONNECT_TIMEOUT_MS = 15_000;
-
-/**
- * How long a finished generation may go on playing before the turn is ended
- * anyway. It is a backstop for a reply that produced no audio at all — or one
- * whose audio drained while its `response.done` never arrived — not the
- * normal path: a spoken reply ends when it goes quiet.
- */
-export const REALTIME_SETTLE_TIMEOUT_MS = 20_000;
-
-/**
- * How long a turn that asked the brain may wait for the answer and the reply
- * voicing it. A brain turn reads transcripts and may act before it answers,
- * so the ordinary settle backstop is far too short for it; the main process's
- * own wait answers a turn the brain has not finished with "still working",
- * and this is the backstop under that.
- */
-export const BRAIN_ASK_SETTLE_TIMEOUT_MS = 60_000;
-
-/**
- * Whose words the caption is showing: a briefing the brain decided to give,
- * or a reply to the developer. History records the two differently.
- */
-export const REPLY_KIND = {
-  BRIEFING: BRIEFING_SPEECH_KIND,
-  REPLY: "reply",
-} as const;
-
-export type ReplyKind = (typeof REPLY_KIND)[keyof typeof REPLY_KIND];
-
-/** Bounds interruption events whose successful requests receive no matching acknowledgement. */
-const MAXIMUM_PENDING_INTERRUPTIONS = 24;
-
-const INTERRUPTION_EVENT_KIND = {
-  CANCELLATION: "cancellation",
-  AUDIO_CLEAR: "audio-clear",
-  TRUNCATION: "truncation",
-} as const;
-
-type InterruptionEventKind = (typeof INTERRUPTION_EVENT_KIND)[keyof typeof INTERRUPTION_EVENT_KIND];
-
-const NO_ACTIVE_RESPONSE_CANCELLATION = /^Cancellation failed:\s*no active response\b/i;
-
-/**
- * The server refusing to trim a reply past its own end. The trim measures how
- * long the reply was audible on a wall clock, which outruns the audio itself
- * when a stop lands at the reply's very end — the words all played, the clock
- * kept counting. A reply refused this way was heard whole, so the record the
- * trim would have corrected is already right.
- */
-const TRUNCATION_PAST_AUDIO_END = /^Audio content of \d+ms is already shorter than\b/i;
-
-/**
- * One press's words on their way through a handshake: the capture reading the
- * device — absent once the press was released — and the buffer holding what
- * it has heard until the data channel can carry it.
- */
-interface PressCaptureState {
-  source: PressCaptureSource | undefined;
-  buffer: PressAudioBuffer;
-}
-
-/**
- * The backstop for a reply whose ending never arrives.
- *
- * `output_audio_buffer.stopped` is what actually ends a reply now, so this only
- * has to catch a call where that never came. It is long because the thing it
- * must not mistake for an ending is a pause between two sentences: at 700ms it
- * did exactly that, taking the meter and the face down while Luke talked on
- * into the second one. The meter itself calls quiet after a fifth of a second,
- * which is shorter still, so a turn that ended on the meter's edge would do
- * the same.
- */
-export const REMOTE_QUIET_MS = 2_500;
-
-/**
- * How many back-to-back responses the caption keeps on screen at once. Two is
- * the shape the surface stacks — the words just settled and the words now
- * arriving — and a third response starting simply retires the oldest, the way
- * a long reply's oldest lines already roll up under the shape.
- */
-export const CAPTION_SEGMENT_LIMIT = 2;
-
-/**
- * Carries one app act the brain decided — a settings change, the panel shown,
- * the feedback composer brought up, the Updates row's button — to the renderer
- * that can perform it, and answers what became of it. The act was validated
- * against the guide in the main process before it got here; the carrier only
- * performs and reports. Nothing here sends a note: the feedback act opens the
- * composer, and what it holds leaves only by its own Send button.
- */
-export type AppActionCarrier = (action: BrainAppActRequest["action"]) => Promise<WireRecord>;
-
-export interface RealtimeVoiceSessionCallbacks {
-  onStatus(status: RealtimeStatus): void;
-  onLocalStream(stream: MediaStream | undefined): void;
-  onRemoteStream(stream: MediaStream | undefined): void;
-  onError(message: string | undefined): void;
-  /**
-   * The words Luke is currently speaking, growing as they are generated, or
-   * undefined once there is nothing being spoken. Each entry is one response's
-   * words: a turn that speaks twice — a sentence before a tool call and the
-   * outcome after it, or a reply the model split into two messages — hands
-   * over both, oldest first, so the surface can stack them apart instead of
-   * running two sentences together. The session owns the whole lifecycle —
-   * the captions clear when the reply ends, is cut off, or the call closes —
-   * so the caller only ever draws what it is handed. `kind` says whether the
-   * words are a briefing or a reply to the brain's answer, living exactly as
-   * long as that reply; a reply the brain was not asked for carries none.
-   */
-  onCaption(texts: readonly string[] | undefined, kind: ReplyKind | undefined): void;
-  /**
-   * The words a reply leaves behind at the moment it ends — finished, talked
-   * over, or the call closing under it, whichever came. `kind` says whether
-   * the words were a briefing or a reply, so History records each as itself.
-   * The words were already spoken toward the room (the caption runs a little
-   * ahead of the audio, so a cut reply hands over slightly more than was
-   * heard); the caller records them so the thread survives the call. A reply
-   * voicing a brain run's end names that run: its words already stand in the
-   * thread, written by the main process from the record, and the caller
-   * records nothing for it. Each reply carries its own, so two overlapping —
-   * one cut off by the next — can never trade attributions.
-   */
-  onReplyEnded?(texts: readonly string[], kind: ReplyKind | undefined, runId?: string): void;
-  /**
-   * The developer's own spoken turn, as the voice service transcribed it. It
-   * arrives on the transcription's clock — often after the reply to it has
-   * already begun — and only from the developer's own call: a speak-only call
-   * offers no microphone, so it has no spoken turns to hand back. The caller
-   * records the words so the thread holds both halves of the exchange.
-   */
-  onSpokenAsk?(transcript: string, itemId: string): void;
-  /**
-   * The developer's spoken words taking shape, one growing piece at a time,
-   * from the same call and under the same microphone guard as the completed
-   * transcript. Preview only: the caller may draw the ask as it is
-   * transcribed, but records nothing until `onSpokenAsk` hands the settled
-   * words over — a turn whose transcription fails previews and then leaves.
-   */
-  onSpokenAskDelta?(itemId: string, delta: string): void;
-  /**
-   * A spoken turn whose transcription the service gave up on. No completed
-   * transcript is coming, so whatever preview the turn's deltas built must
-   * leave rather than stand forever as words still arriving.
-   */
-  onSpokenAskFailed?(itemId: string): void;
-  /** The local audio turn closed, before its commit can be acknowledged. */
-  onSpokenAskClosed?(): void;
-  /** The server item that fixes which current-launch history a spoken turn belongs to. */
-  onSpokenAskCommitted?(itemId: string): void;
-  /**
-   * A reply concluding, words or none. `onReplyEnded` hands over only words
-   * that exist, so a reply the server failed or answered without a transcript
-   * ends without it — and a caller sequencing on endings alone (the
-   * introduction's scripted beats) would wait forever on one. This fires once
-   * per concluded reply, from the same funnel every ending passes, the settle
-   * backstop included.
-   */
-  onReplySettled?(): void;
-}
-
-/**
- * The browser pieces a call runs on, each stated as the members this file
- * touches rather than as its whole `RTC*` type. A real connection satisfies
- * these; the whole browser types do not work the other way, because the
- * thirty-odd members no test can supply would mean the injection point could
- * only ever take the real thing — which is the opposite of why it exists.
- */
-export interface MicrophoneSender {
-  replaceTrack(next: MediaStreamTrack | null): Promise<void>;
-}
-
-export interface RealtimeVoiceSessionOptions extends RealtimeVoiceSessionCallbacks {
-  requestConnection(): Promise<RealtimeConnection | undefined>;
-  /**
-   * Answers the voice's one tool: the developer's words go to the brain under
-   * the tool call's own id as the submission — so a call the service repeats
-   * finds the same run — and what comes back is the reply for the voice to
-   * say, the note that the run is still going, or a bounded refusal. Absent
-   * means the voice can only speak for itself, and every ask is refused.
-   */
-  askBrain?: (question: string, submissionId: string) => Promise<BrainAskResult>;
-  /** The SDK call seam, injectable so the complete transport can be tested without WebRTC. */
-  createSdkTransport?: SdkTransportFactory;
-  /** The full session document used by the SDK; the introduction narrows this to no tools. */
-  sessionConfig?: (model: string) => BuiltRealtimeSessionConfig;
-  /** The existing hidden playback element owned by the renderer. */
-  audioElement?: () => HTMLAudioElement | null;
-  requestMicrophoneStream?: () => Promise<MediaStream>;
-  /**
-   * The local PCM capture a press runs while its call is still connecting.
-   * Injectable on the browser pieces' own terms: the cold-press seam is a
-   * state machine worth testing without a real audio graph.
-   */
-  createPressCapture?: PressCaptureFactory;
-  connectTimeoutMs?: number;
-  /** Injectable so a test can hold the clock a truncate measures against. */
-  now?: () => number;
-  /**
-   * A development tap on the wire itself: every event this call sends and
-   * every event the service answers, as they cross the data channel. The tap
-   * observes and never steers — nothing here reads its result — and the
-   * caller decides per event whether anything leaves it, because the session
-   * outlives the bootstrap that says whether a trace is being written.
-   */
-  onWireEvent?: (direction: TraceDirection, event: WireRecord) => void;
-}
-
-function errorMessage(error: Error): string {
-  return error.message;
-}
-
-/** The developer's words as the voice handed them to the brain, or nothing worth asking. */
-function askQuestion(argumentsJson: string): string | undefined {
-  try {
-    // SAFETY: JSON.parse returns a wire value; the record and string checks are the validation.
-    const parsed = JSON.parse(argumentsJson) as UnparsedWireValue;
-    if (!isRecord(parsed)) return undefined;
-    return text(parsed.question)?.trim().slice(0, maximumTypedAskLength) || undefined;
-  } catch {
-    return undefined;
-  }
-}
-
-/**
- * Whether a quiet stretch is Luke's to answer for.
- *
- * One meter draws both halves of the conversation, so it goes quiet twice for
- * reasons that have nothing to do with Luke: while it lets go of the
- * microphone as a turn is committed, and again in the gap before the first
- * word comes back. Ending a reply on either takes his waveform down seconds
- * after it appeared, while he is still speaking.
- *
- * So silence only counts once the reply is his and something has been heard of
- * it. Nothing here decides when a reply is over — that is the generation being
- * finished as well — only whose silence is being read.
- */
-export function quietIsLukesOwn(input: { status: RealtimeStatus; heardLuke: boolean }): boolean {
-  return input.status === REALTIME_STATUS.RESPONDING && input.heardLuke;
-}
-
-interface SdkToolBatch {
-  responseId: string;
-  callIds: Set<string>;
-  outputCallIds: Set<string>;
-  epoch: number;
-  responseDone: boolean;
-  followUpStarted: boolean;
-}
-
-/**
- * Drives one Realtime conversation over WebRTC.
- *
- * The capture device is the developer's, not the call's: it opens when a
- * press takes a turn and closes when the exchange that press started settles,
- * and nothing else — not connecting, not typing, not an announcement — ever
- * touches it. The SDK negotiates a generated silent track up front, so each
- * turn's fresh microphone track rides the same sender without renegotiating.
- * The press is held as a pending intention while the device
- * opens, exactly as a press that beats the call's handshake is. The close
- * waits for the settle rather than the key coming up because closing a
- * capture device is itself audible on shared hardware — a Bluetooth headset
- * renegotiates its codec — and the key comes up exactly as Luke starts to
- * answer; the track is disabled at that moment, and the device follows in
- * the quiet after the reply. What this buys is that the microphone and its
- * indicator are user-driven — one exchange, opened by one press — and that
- * other audio is never degraded by a device held while nobody is talking.
- */
-export class RealtimeVoiceSession {
-  readonly #options: RealtimeVoiceSessionOptions;
-  #sdkTransport: SdkRealtimeTransport | undefined;
-  #tearingDown = false;
-  #microphone: MediaStreamTrack | undefined;
-  #stream: MediaStream | undefined;
-  /**
-   * The sender the microphone track rides. It outlives the track on purpose:
-   * a released device leaves the sender on the call, silent, which is what
-   * lets the next press attach a fresh track without renegotiating.
-   */
-  #microphoneSender: MicrophoneSender | undefined;
-  #silenceTrack: MediaStreamTrack | undefined;
-  /** The device being reopened, held so two presses cannot open it twice. */
-  #acquiring: Promise<void> | undefined;
-  /**
-   * The connect attempt the state below belongs to, as an identity: refreshed
-   * by every teardown, captured by a device open before its wait, and what a
-   * device arriving late is checked against — a call closed or replaced while
-   * the device was opening keeps a fresh token, so the stale open releases
-   * what it holds instead of adopting it. Mid-connect there is no sender yet
-   * to stand for the call, so the token is what does.
-   */
-  #attempt = Symbol("connect-attempt");
-  /**
-   * The words the press has spoken while its call is still connecting: a local
-   * PCM capture off the press's own device, buffered until the data channel
-   * can carry them as appends. It belongs to exactly one connect attempt —
-   * created only while a press holds a turn, discarded on every path that ends
-   * the attempt — so no press can leave audio behind for a later connection.
-   *
-   * The seam between the captured words and the live WebRTC track is decided
-   * deliberately: the whole of the turn the press opened travels as appends,
-   * and the track joins the sender only when that turn is over. Appends ride
-   * the ordered data channel while the track rides RTP, and the server writes
-   * one input buffer in arrival order — so any turn that mixed the two would
-   * have an unorderable seam where a late append lands after the first live
-   * frames and words swap, double, or drop. On one channel every chunk lands
-   * exactly once, in capture order, with the commit behind the last of them.
-   * Handing over between turns is safe because every turn opens by clearing
-   * the buffer: there is nothing across that seam to double.
-   */
-  #pressCapture: PressCaptureState | undefined;
-  /**
-   * A press released while the call was still connecting, with words already
-   * captured. The turn it held is over — the capture stops reading and the
-   * device closes at the release — but what it heard is owed a delivery:
-   * flushed and committed once this attempt's channel opens, and discarded
-   * with the attempt if it never does.
-   */
-  #pressCommitPending = false;
-  /** Whether the turn now under way travels as appends rather than the track. */
-  #listeningOnAppends = false;
-  /**
-   * Whether the current connect attempt — and the call it opens — includes the
-   * microphone. A developer's call does; one Luke opens for himself, to read a
-   * notice out, must not: no capture device is held, no permission is asked,
-   * and the macOS microphone indicator stays honest by never lighting for a
-   * call nobody spoke into.
-   */
-  #withMicrophone = true;
-  #status: RealtimeStatus = REALTIME_STATUS.IDLE;
-  #connecting: Promise<boolean> | undefined;
-  #closed = false;
-  /**
-   * Cancel, clear, and trim requests not yet answered with an error, by their
-   * stamped names. Their errors belong to the reply that was interrupted, so
-   * they must never finish a newer turn; only the redundant-cancel race and a
-   * trim refused for asking past the audio's end stay quiet, while every
-   * genuine refusal is still reported.
-   */
-  #pendingInterruptions = new Map<string, InterruptionEventKind>();
-  #interruptionSequence = 0;
-  /**
-   * Luke's own audio track. Cancelling stops the model producing more, but what
-   * it already produced is on its way down the connection and keeps playing —
-   * so cutting him off means silencing this end too, which is the only half of
-   * it entirely under our control.
-   */
-  #remoteTrack: MediaStreamTrack | undefined;
-  /**
-   * Whether the model has finished producing the reply. It is not the same as
-   * the reply being over: `response.done` says generation is complete, and the
-   * audio it produced is still on its way out. A turn that ended here would
-   * take the meter and the face down while Luke was still audible, and would
-   * let the next press start a turn over the top of him.
-   */
-  #generationDone = false;
-  /**
-   * Whether Luke has been quiet since he was last heard. Generation finishing
-   * and playback finishing are two events with no fixed order, and only one of
-   * them arrives twice: the meter reports an edge, so a quiet that lands before
-   * `response.done` is the only quiet there will be. Remembering it is what
-   * lets the second of the two end the turn, whichever one that turns out to
-   * be.
-   */
-  #remoteQuiet = false;
-  /**
-   * Whether the server still owes this turn a `response.done`: raised when a
-   * reply is asked for, lowered when the server concludes it — its own
-   * `done`, the error that refused it outright, or the cancel an interrupt
-   * sends ahead of anything newer. The client's side of the turn can settle
-   * first — the audio drains before the `done` arrives — and in that window
-   * the conversation still holds an active response: a `response.create`
-   * sent into it is refused as a conversation already in progress, with the
-   * refusal read out to the developer as a voice error and the reply it was
-   * meant to open lost. Nothing may end the turn while this stands.
-   */
-  #responseOutstanding = false;
-  /**
-   * The reply whose audio ran out while the server still owed the turn its
-   * `done`, or false while audio is still owed. The ending the drain would
-   * have made is remembered here and lands when the `done` arrives, so the
-   * turn still closes on the second of the two events whichever order they
-   * come in. The drain keeps the name of the response the server said
-   * drained, because it can be an old reply's arriving late — a tool turn's
-   * spoken half empties after its follow-up was already asked for — and a
-   * stale drain read as the current reply's would end the follow-up under
-   * Luke's own voice, or skip the trim a stop mid-follow-up still owes the
-   * record.
-   */
-  #audioDrained: { responseId: string | undefined } | false = false;
-  /**
-   * Whether an armed reply's calls are being answered and the follow-up
-   * voicing their outcomes is still owed. The turn holds through the writes
-   * — a READY offered mid-write is the edge the announcer rides, and a reply
-   * taken there bumps the epoch and abandons the follow-up — so the audio
-   * draining then is remembered rather than an ending. Lowered by whatever
-   * reply comes next, the finish that ends the hold, or the interrupt of a
-   * developer moving on.
-   */
-  #toolBatch: SdkToolBatch | undefined;
-  #toolCallResponseIds = new Map<string, string>();
-  /**
-   * Whether Luke has actually been heard during this reply. Committing a turn
-   * swaps the meter from the microphone to Luke, and the meter reports quiet as
-   * it lets go of the old stream — a silence that belongs to the developer, not
-   * to Luke, and one that would otherwise end his turn before he had said
-   * anything.
-   */
-  #heardLuke = false;
-  /**
-   * The pause between two sentences is longer than the meter's idea of quiet.
-   * This holds that pause until it has lasted longer than speech leaves behind.
-   */
-  #quietTimer: ReturnType<typeof setTimeout> | undefined;
-  /**
-   * A press of the talk key that arrived before there was a call to press
-   * against. The microphone opens only once the call is up, so such a press is
-   * an intention rather than a turn.
-   */
-  #pendingTurn = false;
-  /**
-   * The message Luke's current reply is being spoken into, and the moment it
-   * first became audible. Together they are what a truncate needs: which
-   * message to cut, and how much of it reached the room.
-   */
-  #responseItemId: string | undefined;
-  /**
-   * The response now under way, as the server named it when it confirmed the
-   * reply had started — or nothing between asking for a reply and that
-   * confirmation. It is what tells the current reply's `response.done` from a
-   * cancelled one's: the server had finished composing the old reply before
-   * the cancel landed, so its `done` still arrives, and it can carry tool
-   * calls. Matching the id is what keeps those calls from being answered with
-   * the new turn's arming — the turn that superseded them, not the one that
-   * asked.
-   */
-  #activeResponseId: string | undefined;
-  #audibleSince: number | undefined;
-  /**
-   * The words of the turn being spoken, as far as they have arrived — one
-   * segment per output item, oldest first, each remembering which item spoke
-   * it. A turn that speaks twice back-to-back — a second message item, or the
-   * follow-up after a tool call — starts a new segment rather than running
-   * its words onto the last one's, and an item's own final transcript can
-   * still land on its own segment after the turn has moved on. Kept here
-   * rather than in the caller so every path that ends a reply — finishing,
-   * being talked over, the call dropping — clears the words with it, and a
-   * caption can never outlive the speech it captions.
-   */
-  #captionSegments: { itemId: string | undefined; text: string }[] = [];
-  /**
-   * Whether the words under way are a briefing or a reply, for History to
-   * record as such. Set only when the words were decided by the brain and
-   * cleared wherever the caption is, so it can never outlive the reply.
-   */
-  #captionKind: ReplyKind | undefined;
-  /**
-   * The brain run whose end the reply under way is voicing, when it is one.
-   * Set with the kind and cleared wherever the caption is, so it is exactly
-   * as long-lived as the reply it names and no other reply can inherit it.
-   */
-  #captionRunId: string | undefined;
-  /**
-   * Whether this call has ever reported a reply's audio running out. Once it
-   * has, silence stops being evidence of anything: the server says when Luke is
-   * finished, and a stretch of quiet is just as likely to be the gap between
-   * two sentences. Calls that never report one keep the old guess, because a
-   * turn that never ends is worse than one that ends early.
-   */
-  #audioEndingsReported = false;
-  #settleTimer: ReturnType<typeof setTimeout> | undefined;
-  /**
-   * A monotonic id for the turn now under way, bumped at every boundary a
-   * turn crosses — a new one beginning, or the old one declared over. A tool
-   * follow-up captures it before awaiting the write and refuses to open if it
-   * has changed — the developer has taken the turn, started another, or the
-   * turn ended without it, the settle backstop giving up on a hung write —
-   * so Luke never speaks the outcome over a live microphone, over a reply the
-   * developer is already hearing, or out of a silence already declared.
-   */
-  #turnEpoch = 0;
-  /**
-   * A pace change that arrived mid-reply, waiting for the reply to end. The
-   * API applies a speed only between model turns, so one landing while Luke is
-   * speaking is held here and sent ahead of whatever the call does next.
-   */
-  #pendingSpeed: number | undefined;
-  constructor(options: RealtimeVoiceSessionOptions) {
-    this.#options = options;
-  }
-
-  get status(): RealtimeStatus {
-    return this.#status;
-  }
-
-  get isConnected(): boolean {
-    return this.#sdkTransport?.status === "connected";
-  }
-
-  /**
-   * Whether the call that is up — or coming up — is one the developer can take
-   * a turn on. A call Luke opened to read a notice out has no microphone and
-   * answers false, which is how a talk-key press knows it still has a call of
-   * its own to open. The developer's call whose device is resting between
-   * turns still answers true: the call can take the turn, and the device
-   * rejoins it on the press rather than a fresh call replacing it.
-   */
-  get microphoneCall(): boolean {
-    if (this.isConnected) return this.#withMicrophone;
-    if (this.isConnecting) return this.#withMicrophone;
-    return false;
-  }
-
-  /**
-   * Opens the call, reusing an in-flight attempt rather than racing a second
-   * one. Without the microphone it is a call Luke opens for himself — audio
-   * out only, nothing captured, nothing to consent to — and it always stands
-   * down for the developer's own: asking for a microphone call while a
-   * speak-only one is up replaces it rather than sharing it.
-   */
-  async connect(options?: { microphone?: boolean }): Promise<boolean> {
-    const withMicrophone = options?.microphone !== false;
-    // Wait out whatever attempt is already in flight rather than racing it.
-    // A loop, because another caller can start a new attempt in the gap.
-    while (this.#connecting) await this.#connecting;
-    if (this.isConnected) {
-      if (!withMicrophone || this.#microphone) return true;
-      // The developer's call replaces Luke's own — but the press that asked
-      // for it must survive the teardown, or the turn it opens would be lost.
-      const pendingTurn = this.#pendingTurn;
-      this.#teardown();
-      this.#pendingTurn = pendingTurn;
-    }
-    this.#withMicrophone = withMicrophone;
-    this.#connecting = this.#connect()
-      .then((opened) => {
-        // A press does not outlive the attempt it started. Every way a connect
-        // ends without a call passes through here — no credential, a refused
-        // call, a timeout — so none of them can leave an intention behind for
-        // some later connection to open a turn nobody asked for. The words a
-        // press captured are under the same rule: the teardown each of those
-        // paths runs has already discarded them, and this clears the delivery
-        // they were owed.
-        if (!opened) {
-          this.#pendingTurn = false;
-          this.#pressCommitPending = false;
-        }
-        return opened;
-      })
-      .finally(() => {
-        this.#connecting = undefined;
-      });
-    return this.#connecting;
-  }
-
-  async #connect(): Promise<boolean> {
-    this.#closed = false;
-    this.#setStatus(REALTIME_STATUS.CONNECTING);
-    this.#options.onError(undefined);
-
-    // Connecting opens no capture device of its own: the microphone is the
-    // developer's, not the call's, and it opens only when a press takes a
-    // turn. A press that started this connect has already asked for it, and
-    // its capture rides beside the mint — but that is the press's doing,
-    // never the connect's.
-    let connection: RealtimeConnection | undefined;
-    try {
-      connection = await this.#options.requestConnection();
-    } catch (error) {
-      // A stop that lands while the credential is being minted is not a fault
-      // to report. Every exit from here has to ask, not just the ones after
-      // the handshake starts.
-      if (this.#closed) return this.#abandonConnect();
-      if (!(error instanceof Error)) return this.#fail(String(error));
-      return this.#fail(`Could not reach the main process: ${errorMessage(error)}`);
-    }
-    if (this.#closed) return this.#abandonConnect();
-    if (!connection) {
-      this.#teardown();
-      this.#setStatus(REALTIME_STATUS.UNAVAILABLE);
-      return false;
-    }
-
-    try {
-      const factory = this.#options.createSdkTransport ?? createAgentsRealtimeTransport;
-      const sdkTransport = factory({
-        sessionConfig: (
-          this.#options.sessionConfig ?? ((model) => realtimeSessionConfig({ model }))
-        )(connection.model),
-        audioElement: this.#options.audioElement?.() ?? undefined,
-        onPeerConnection: (peer, silenceTrack) => this.#adoptPeerConnection(peer, silenceTrack),
-        onTransportEvent: (event) => this.#handleServerEvent(event),
-        onClientEvent: (event) => this.#options.onWireEvent?.(TRACE_DIRECTION.CLIENT, event),
-        onToolOutputSent: (callId) => this.#toolOutputSent(callId),
-        onConnectionChange: (status) => this.#connectionChanged(status),
-        onError: (message) => this.#options.onError(message),
-        executeTool: (name, details) => this.#sdkToolOutput(name, details),
-      });
-      this.#sdkTransport = sdkTransport;
-      const timeoutMs = positiveInteger(this.#options.connectTimeoutMs, CONNECT_TIMEOUT_MS);
-      let deadlineTimer: ReturnType<typeof setTimeout> | undefined;
-      const deadline = new Promise<never>((_resolve, reject) => {
-        deadlineTimer = setTimeout(() => {
-          reject(new Error("The voice connection timed out while opening."));
-        }, timeoutMs);
-      });
-      try {
-        await Promise.race([
-          sdkTransport.connect({
-            apiKey: connection.value,
-            model: connection.model,
-            url: connection.callsUrl,
-          }),
-          deadline,
-        ]);
-      } finally {
-        if (deadlineTimer !== undefined) clearTimeout(deadlineTimer);
-      }
-      if (this.#closed) return this.#abandonConnect();
-      this.#setStatus(REALTIME_STATUS.READY);
-      // A pace changed during the handshake could not be sent then, and the
-      // credential this call answered may have been minted before the change.
-      this.#flushPendingSpeed();
-      // Whoever pressed the talk key to get here has been waiting through the
-      // handshake for their turn to open. The press already opened the device
-      // and has been captured since it answered, so the turn opens on those
-      // words — as appends, with the track joining the sender only when the
-      // turn is over. A press whose device is still opening falls back to the
-      // acquire, and its turn opens when the device arrives. A speak-only
-      // call leaves the press pending instead — it has no sending half to
-      // open a turn with, and the developer's own call is already replacing
-      // it.
-      if (this.#pendingTurn && this.#withMicrophone) {
-        if (this.#pressCapture && this.#microphone) {
-          this.#pendingTurn = false;
-          this.#beginAppendsTurn();
-        } else {
-          this.#acquireMicrophone();
-        }
-      } else if (this.#pressCommitPending) {
-        // The press was released mid-handshake. Its words go as the turn it
-        // held — one tick later, because the caller re-feeds the roster and
-        // the guide right after this connect resolves, and the reply to those
-        // words must be answered from that context rather than from none.
-        setTimeout(() => this.#deliverHeldTurn(), 0);
-      }
-      return true;
-    } catch (error) {
-      if (this.#closed) return this.#abandonConnect();
-      if (!(error instanceof Error)) return this.#fail(String(error));
-      return this.#fail(errorMessage(error));
-    }
-  }
-
-  /**
-   * Releases everything a cancelled connect had already acquired, leaving the
-   * session idle as the caller that stopped it intended.
-   */
-  #abandonConnect(): boolean {
-    this.#teardown();
-    this.#setStatus(REALTIME_STATUS.IDLE);
-    return false;
-  }
-
-  /**
-   * How the device is asked for — only ever on behalf of a press. The caller
-   * usually injects the route-aware opener; the fallback opens the browser's
-   * default with the same processing (`MICROPHONE_PROCESSING` says why echo
-   * cancellation is off), so a bare session still captures correctly.
-   */
-  #requestStream(): Promise<MediaStream> {
-    return (
-      this.#options.requestMicrophoneStream?.() ??
-      navigator.mediaDevices.getUserMedia({ audio: { ...MICROPHONE_PROCESSING }, video: false })
-    );
-  }
-
-  #adoptPeerConnection(peer: RTCPeerConnection, silenceTrack: MediaStreamTrack): void {
-    const sender = peer.getSenders().find((candidate) => candidate.track === silenceTrack);
-    if (!sender) throw new Error("The voice connection did not provide an audio sender.");
-    this.#silenceTrack = silenceTrack;
-    this.#microphoneSender = sender;
-    const sdkOnTrack = peer.ontrack;
-    peer.ontrack = (event) => {
-      sdkOnTrack?.call(peer, event);
-      this.#remoteTrack = event.track;
-      this.#options.onRemoteStream(event.streams[0] ?? new MediaStream([event.track]));
-    };
-    // The listener outlives the peer it was added to, so a stale peer's late
-    // `failed` is answered by the attempt token rather than by the live call.
-    const attempt = this.#attempt;
-    peer.addEventListener("connectionstatechange", () => {
-      if (attempt !== this.#attempt || this.#closed) return;
-      if (peer.connectionState === "failed") this.#fail("The voice connection dropped.");
-    });
-  }
-
-  #connectionChanged(status: SdkRealtimeTransport["status"]): void {
-    // Teardown clears the transport before anything can report on it, so a
-    // `disconnected` with no transport standing is a call already ended — by
-    // a stop or a failure whose status must not be rewritten to idle.
-    if (
-      status !== "disconnected" ||
-      this.#closed ||
-      this.#tearingDown ||
-      !this.#sdkTransport ||
-      this.#status === REALTIME_STATUS.CONNECTING
-    ) {
-      return;
-    }
-    if (this.#captionKind !== undefined) this.#discardCaption();
-    this.#teardown();
-    this.#setStatus(REALTIME_STATUS.IDLE);
-  }
-
-  /**
-   * The talk key going down. Opening a turn and ending one are separate here
-   * rather than one toggle, because a key that reports being let go of can say
-   * which of the two it meant — and a turn that lasts exactly as long as the
-   * key is held cannot be left open by forgetting to press again.
-   */
-  beginTurn(): void {
-    // Every press opens the device: it lives exactly as long as the turn it
-    // was pressed for. Until the device is live — and, on a first press, the
-    // call it rides — the press waits as an intention and the turn opens the
-    // moment both exist.
-    if (!this.isConnected || !this.#microphone) {
-      // The developer's turn wins at the press, not at the device: a reply
-      // under way is cut off now, exactly as the stop key cuts one, and the
-      // turn itself opens when the device arrives.
-      this.stopSpeaking();
-      this.#pendingTurn = true;
-      this.#acquireMicrophone();
-      return;
-    }
-    this.startListening();
-  }
-
-  /**
-   * The talk key coming up on a held turn, or a second press ending a latched
-   * one. A turn that never opened is dropped rather than committed: the
-   * microphone opens with the press, and one that was let go of before the
-   * device arrived captured nothing to send.
-   */
-  endTurn(commit: boolean): void {
-    if (!this.isConnected) {
-      this.#pendingTurn = false;
-      const capture = this.#pressCapture;
-      if (commit && capture && !capture.buffer.isEmpty) {
-        // The press already spoke, so its words are owed a delivery once this
-        // attempt's channel opens. The press no longer holds a turn, though:
-        // the capture stops reading and the device closes this instant — the
-        // sealed words wait in memory, not on an open microphone.
-        capture.source?.stop();
-        capture.source = undefined;
-        this.#pressCommitPending = true;
-        this.#releaseMicrophone();
-        return;
-      }
-      // Nothing was captured toward this press — the device never arrived,
-      // or nothing was said into it — so the turn it was owed is dropped, as
-      // committing it would ask the server to answer an empty buffer.
-      this.#retirePressCapture();
-      this.#releaseMicrophone();
-      return;
-    }
-    // A press let go of while its device was still opening held nothing of
-    // its own to send: the turn it was owed is dropped rather than opened
-    // under a key that is already up. But a re-press re-opens a sealed turn,
-    // and the words sealed toward it are still owed — a commit delivers them
-    // now, connected as we are, and a discard lets the whole turn go. A press
-    // against Luke's speak-only call is a different wait — for the
-    // developer's call — and keeps its meaning.
-    if (!this.#microphone && this.#withMicrophone) {
-      this.#pendingTurn = false;
-      if (commit && this.#pressCommitPending) this.#deliverHeldTurn();
-      else this.#retirePressCapture();
-      return;
-    }
-    this.stopListening(commit);
-  }
-
-  /** Whether a call is being opened, so a press has something to wait for. */
-  get isConnecting(): boolean {
-    return this.#connecting !== undefined;
-  }
-
-  /**
-   * Whether a press is still waiting for a call that can take its turn. The
-   * opening meter reads this: a takeover — the developer's call replacing
-   * Luke's own — passes through a settled status on the way, and the meter
-   * must not come down while the press that started it is still owed a turn.
-   * A press released mid-connect with words captured is owed one too — its
-   * delivery — so the meter rides until the reply to it begins.
-   */
-  get turnPending(): boolean {
-    return this.#pendingTurn || this.#pressCommitPending;
-  }
-
-  /**
-   * Forgets a press that was waiting for a call that is not coming — a refused
-   * microphone, say. Without this the intention would outlive the attempt and
-   * open a turn out of the next connection, which nobody asked for.
-   */
-  dropPendingTurn(): void {
-    this.#pendingTurn = false;
-    // The words captured toward the dropped press go with it.
-    this.#retirePressCapture();
-    // A device already opened for that press has no turn left to serve, and
-    // nobody is talking into it: it closes now, not on any clock.
-    if (this.#status !== REALTIME_STATUS.LISTENING) this.#releaseMicrophone();
-  }
-
-  /**
-   * Opens the capture device for the press waiting on it. One request at a
-   * time: a second press mid-open joins the first rather than racing it. A
-   * press ahead of the call opens the device too — the capture beside the
-   * handshake is what carries the words spoken into it — so being connected
-   * is not required, only being a call that will take a microphone.
-   */
-  #acquireMicrophone(): void {
-    if (!this.#withMicrophone || this.#microphone) return;
-    if (this.#acquiring) return;
-    this.#acquiring = this.#openMicrophone().finally(() => {
-      this.#acquiring = undefined;
-      // A press that arrived while a stale open was still in flight was told
-      // to wait by the guard above; with the flight over, it is served now.
-      if (this.#pendingTurn && !this.#microphone) this.#acquireMicrophone();
-    });
-  }
-
-  async #openMicrophone(): Promise<void> {
-    // The attempt names the call — or the connect under way — this open
-    // belongs to, captured before the wait: one closed and replaced while the
-    // device was opening keeps a fresh token, which is how this open knows it
-    // has gone stale. The sender cannot stand for the call here, because a
-    // press ahead of the handshake opens the device before any sender exists.
-    const attempt = this.#attempt;
-    let stream: MediaStream;
-    try {
-      stream = await this.#requestStream();
-    } catch (error) {
-      // A refusal belongs to the attempt whose press asked for the device. If
-      // that attempt is gone — closed, or replaced mid-open — the refusal
-      // died with it, and the call now up must not be torn down for it. On a
-      // call already standing it is failed rather than left looking able to
-      // listen, and `FAILED` offers "Start voice" again; mid-connect the call
-      // the press is still waiting for goes on opening, the press is dropped —
-      // there is nothing to capture with — and the refusal is shown beside it.
-      if (this.#closed || attempt !== this.#attempt) return;
-      const message = error instanceof Error ? errorMessage(error) : String(error);
-      if (this.isConnected) {
-        this.#fail(message);
-        return;
-      }
-      this.#pendingTurn = false;
-      this.#options.onError(message);
-      return;
-    }
-    // The attempt may have gone away while the device was opening. A device
-    // nobody adopts is released here, or it would hold the indicator lit with
-    // nothing left to close it.
-    if (this.#closed || attempt !== this.#attempt || this.#microphone) {
-      for (const track of stream.getTracks()) track.stop();
-      return;
-    }
-    const [microphone] = stream.getAudioTracks();
-    if (!microphone) {
-      for (const track of stream.getTracks()) track.stop();
-      if (this.isConnected) {
-        this.#fail("No microphone track was available.");
-      } else {
-        this.#pendingTurn = false;
-        this.#options.onError("No microphone track was available.");
-      }
-      return;
-    }
-    // Closed until the turn opens it, so nothing is sent before the state
-    // machine says the turn is on.
-    microphone.enabled = false;
-    this.#stream = stream;
-    this.#microphone = microphone;
-    this.#options.onLocalStream(stream);
-    if (!this.isConnected) {
-      // The call is still connecting, and the press is waiting it out: the
-      // words start being captured now, and the turn they open travels as
-      // appends — the track joins the sender only when that turn is over. A
-      // press already let go of leaves nothing to capture for, and a connect
-      // that turned speak-only under the press takes no device at all: either
-      // way the device closes as fast as it arrived.
-      if (this.#pendingTurn && this.#withMicrophone) this.#beginPressCapture();
-      else this.#releaseMicrophone();
-      return;
-    }
-    if (this.#pendingTurn && this.#pressCapture) {
-      // The device arrived connected, for a re-press over sealed words — the
-      // channel opened while it was still on its way. The turn it re-opened
-      // still owes those words, and a track turn would start by clearing
-      // them: so the turn opens as the captured turn it began as, resuming
-      // capture on the device that just arrived, and the track joins the
-      // sender at the seam as every captured turn's does.
-      this.#beginPressCapture();
-      this.#pendingTurn = false;
-      this.#beginAppendsTurn();
-      return;
-    }
-    const sender = this.#microphoneSender;
-    if (!sender) {
-      // A connected microphone call always negotiated a sender; without one
-      // there is nothing for a turn to ride, and the device has no taker.
-      this.#releaseMicrophone();
-      return;
-    }
-    try {
-      // Onto the sender the call has kept since its handshake: no
-      // renegotiation, the same line, the same call.
-      await sender.replaceTrack(microphone);
-    } catch (error) {
-      // Guarded like the open's own refusal: a sender that rejected because
-      // its call was torn down mid-replace is not the live call's fault.
-      if (!this.#closed && attempt === this.#attempt && this.isConnected) {
-        if (error instanceof Error) this.#fail(errorMessage(error));
-        else this.#fail(String(error));
-      }
-      return;
-    }
-    if (this.#closed || !this.isConnected) return;
-    if (this.#pendingTurn) {
-      this.#pendingTurn = false;
-      this.startListening();
-      return;
-    }
-    // Opened for a press that has since been let go: nobody is talking, so
-    // the device closes as fast as it arrived.
-    this.#releaseMicrophone();
-  }
-
-  /**
-   * Puts the capture device away without touching the call: the tracks stop,
-   * the sender stays to take the next track, and the meter lets go of a
-   * stream that no longer exists. From here the microphone indicator is dark
-   * and Bluetooth audio is back on its music codec.
-   */
-  #releaseMicrophone(): void {
-    if (!this.#stream && !this.#microphone) return;
-    const stream = this.#stream;
-    this.#stream = undefined;
-    this.#microphone = undefined;
-    stream?.getTracks().forEach((track) => {
-      track.stop();
-    });
-    void this.#microphoneSender?.replaceTrack(this.#silenceTrack ?? null);
-    this.#options.onLocalStream(undefined);
-  }
-
-  /**
-   * Starts capturing the press's words, when there is a press to capture for.
-   * Harmless to ask again: it runs only while a press holds a turn on a
-   * microphone call with the device already open and nothing already reading
-   * it — every other moment it does nothing. Its two callers are the device
-   * arriving mid-connect, and the device arriving connected for a re-press
-   * over sealed words: either way the turn it feeds travels as appends.
-   */
-  #beginPressCapture(): void {
-    if (!this.#withMicrophone || !this.#pendingTurn) return;
-    if (this.#pressCapture?.source || this.#closed) return;
-    const stream = this.#stream;
-    const microphone = this.#microphone;
-    if (!stream || !microphone) return;
-    // A press landing again over words a release already sealed re-opens the
-    // same turn: the delivery the release was owed is superseded — this
-    // press's own release decides afresh — and capture resumes into the same
-    // buffer, so neither press's words are lost.
-    this.#pressCommitPending = false;
-    const capture: PressCaptureState = this.#pressCapture ?? {
-      source: undefined,
-      buffer: new PressAudioBuffer(),
-    };
-    const buffer = capture.buffer;
-    this.#pressCapture = capture;
-    capture.source = (this.#options.createPressCapture ?? createPressCaptureSource)(
-      stream,
-      (chunk) => {
-        // A chunk from a capture this session has already let go of belongs
-        // to no turn and goes nowhere.
-        if (this.#pressCapture !== capture) return;
-        if (this.#listeningOnAppends) {
-          this.#send(inputAudioAppendEvents(chunk));
-          return;
-        }
-        buffer.push(chunk);
-      },
-    );
-    // The capture reads the track, and a disabled track reads as silence to
-    // every consumer — so the track is open exactly while the capture is. The
-    // sender carries no track yet, so nothing reaches the network.
-    microphone.enabled = true;
-  }
-
-  /**
-   * Ends the press capture, discarding whatever it still holds, and hands the
-   * device's track to the sender so the turns after the seam ride WebRTC as
-   * every turn did before it. Every path out of the captured-turn machinery
-   * ends here — the seam settling, a discarded press, a failed attempt — so
-   * none of them can leave the capture reading the device.
-   */
-  #retirePressCapture(): void {
-    const capture = this.#pressCapture;
-    this.#pressCapture = undefined;
-    this.#listeningOnAppends = false;
-    this.#pressCommitPending = false;
-    capture?.source?.stop();
-    // The track was open for the capture to read; nothing reads it now, so it
-    // closes until a turn opens it.
-    if (capture?.source && this.#microphone) this.#microphone.enabled = false;
-    if (capture && this.isConnected && this.#microphoneSender && this.#microphone) {
-      void this.#microphoneSender.replaceTrack(this.#microphone).catch(() => undefined);
-    }
-  }
-
-  /**
-   * Opens the turn a still-held press has been capturing: `startListening` in
-   * every respect but the transport. The words captured so far flush behind a
-   * clean buffer, and the capture keeps appending live from here — the track
-   * stays off the sender, because the whole of this turn travels on the one
-   * ordered channel.
-   */
-  #beginAppendsTurn(): void {
-    const capture = this.#pressCapture;
-    if (!capture || !this.#microphone) {
-      this.#acquireMicrophone();
-      return;
-    }
-    this.#flushHeldAudio(capture);
-    this.#listeningOnAppends = true;
-    this.#turnEpoch += 1;
-    this.#setStatus(REALTIME_STATUS.LISTENING);
-  }
-
-  /**
-   * Sends one captured turn's opening, in the one order the service accepts
-   * it: the format the appends are to be read as, then a clear, then the
-   * held audio. Appends ahead of either are read against whatever the last
-   * turn left behind.
-   */
-  #flushHeldAudio(capture: PressCaptureState): void {
-    this.#send(inputAudioFormatUpdateEvents());
-    this.#send(clearInputAudioEvents());
-    for (const chunk of capture.buffer.drain()) {
-      this.#send(inputAudioAppendEvents(chunk));
-    }
-  }
-
-  /**
-   * Delivers the turn a press held and released while the call was still
-   * connecting: the captured words flush as appends and commit as the turn
-   * the release already closed. It runs a tick after the channel opened so
-   * the caller's context re-feed lands first, and it yields to anything that
-   * moved in that tick — a new turn is the developer talking again, and words
-   * that yielded are discarded rather than queued behind it.
-   */
-  #deliverHeldTurn(): void {
-    const capture = this.#pressCapture;
-    if (!capture || !this.#pressCommitPending) return;
-    if (this.#closed || !this.isConnected || voiceExchangeActive(this.#status)) {
-      this.#retirePressCapture();
-      return;
-    }
-    this.#pressCommitPending = false;
-    this.#flushHeldAudio(capture);
-    this.#retirePressCapture();
-    // A turn exactly as a live commit is: the developer opened it by holding
-    // the key and spoke into it; only the delivery waited.
-    this.#startResponse(pushToTalkCommitEvents());
-  }
-
-  /**
-   * Opens the microphone for as long as push-to-talk is held, reporting
-   * whether it actually did. The caller uses that to decide whether to claim
-   * the key it was pressed with — Space still scrolls the panel when there is
-   * no turn to open.
-   */
-  startListening(): boolean {
-    if (!this.#microphone || !this.isConnected) return false;
-    if (this.#status === REALTIME_STATUS.LISTENING) return false;
-    // Talking over Luke stops it. The developer's turn always wins, which is
-    // the whole point of a key that means "it is my turn now".
-    if (this.#status === REALTIME_STATUS.RESPONDING) this.#interruptReply();
-    // Start from an empty buffer: a muted track still transmits, and with turn
-    // detection off the server keeps everything since the last commit.
-    this.#send(clearInputAudioEvents());
-    this.#microphone.enabled = true;
-    // The developer taking the turn is a new turn, whatever a tool follow-up
-    // still in flight from the last one thinks: it will find this epoch and
-    // stand down rather than talk over the microphone now opening.
-    this.#turnEpoch += 1;
-    this.#setStatus(REALTIME_STATUS.LISTENING);
-    return true;
-  }
-
-  /**
-   * Closes the microphone and either asks for a reply or discards the turn.
-   * Discarding matters: a press the developer changes their mind about must not
-   * leave buffered audio behind for the next turn to inherit.
-   */
-  stopListening(commit: boolean): void {
-    if (!this.#microphone || this.#status !== REALTIME_STATUS.LISTENING) return;
-    if (this.#listeningOnAppends) {
-      // The captured turn ends here whichever way, and the seam settles with
-      // it: the capture stops and the track joins the sender, so every turn
-      // after this one rides WebRTC as before. The device itself is kept
-      // exactly as a track turn keeps it — through the reply, released in
-      // the quiet after it — for the same shared-hardware reason.
-      this.#retirePressCapture();
-      this.#microphone.enabled = false;
-      if (!commit) {
-        this.#send(clearInputAudioEvents());
-        this.#setStatus(REALTIME_STATUS.READY);
-        return;
-      }
-      // The commit follows the last append on the same ordered channel, so it
-      // closes over every word the capture delivered.
-      this.#startResponse(pushToTalkCommitEvents());
-      return;
-    }
-    this.#microphone.enabled = false;
-    if (!commit) {
-      this.#send(clearInputAudioEvents());
-      // Settling to READY is what releases the device: nothing is coming
-      // that its closing could talk over.
-      this.#setStatus(REALTIME_STATUS.READY);
-      return;
-    }
-    // The developer opened this turn and spoke into it, so an ask the voice
-    // makes of the brain out of it is the developer's own. The device is NOT
-    // released here, deliberately: closing a capture device is itself audible
-    // on shared hardware — a Bluetooth headset renegotiates its codec and
-    // playback drops out for a beat — and this is the very moment Luke starts
-    // to answer. The track is disabled, so nothing is sent; the device itself
-    // is let go when the exchange settles, in the quiet after the reply.
-    this.#startResponse(pushToTalkCommitEvents());
-  }
-
-  /**
-   * Speaks the brain's reply to a typed ask, reporting whether it could. The
-   * ask itself went to the brain over the bridge — the voice never saw it —
-   * so what the call is handed is the finished reply, on the briefing's own
-   * out-of-band terms. A reply arriving over another interrupts it: the
-   * developer's turn always wins, however it is taken.
-   */
-  speakReply(briefing: string, runId?: string): boolean {
-    if (!this.isConnected || !this.#withMicrophone) return false;
-    if (this.#status === REALTIME_STATUS.LISTENING) return false;
-    const events = briefingSpeechEvents({
-      kind: BRIEFING_SPEECH_KIND,
-      briefing,
-      decidedAt: Date.now(),
-    });
-    if (events.length === 0) return false;
-    if (this.#status === REALTIME_STATUS.RESPONDING) this.#interruptReply();
-    this.#startResponse(events);
-    this.#captionKind = REPLY_KIND.REPLY;
-    this.#captionRunId = runId;
-    this.#emitCaption();
-    return true;
-  }
-
-  /**
-   * Cuts off the reply under way so the developer's new turn does not land on
-   * top of it.
-   */
-  #interruptReply(): void {
-    // Stop the words that are already on their way, then stop more being
-    // made. A disabled track drops what is buffered rather than playing it
-    // out, so the cut-off is immediate rather than eventual.
-    this.#silenceLuke();
-    // The caption is cut with the audio, but handed over first. Generated text
-    // runs slightly ahead of playback; History keeps that available transcript
-    // so an interrupted announcement can still be recalled.
-    this.#clearCaption();
-    this.#interruptionSequence += 1;
-    const cancellationEventId = `response_cancel_${this.#interruptionSequence}`;
-    const clearEventId = `output_audio_clear_${this.#interruptionSequence}`;
-    const truncationEventId = `item_truncate_${this.#interruptionSequence}`;
-    const truncateEvents = this.#truncateEvents(truncationEventId);
-    const cancelGeneration = this.#responseOutstanding;
-    const interruptionCount = (cancelGeneration ? 2 : 1) + (truncateEvents.length > 0 ? 1 : 0);
-    while (this.#pendingInterruptions.size + interruptionCount > MAXIMUM_PENDING_INTERRUPTIONS) {
-      const [oldest] = this.#pendingInterruptions.keys();
-      if (oldest === undefined) break;
-      this.#pendingInterruptions.delete(oldest);
-    }
-    if (cancelGeneration) {
-      this.#pendingInterruptions.set(cancellationEventId, INTERRUPTION_EVENT_KIND.CANCELLATION);
-      this.#pendingInterruptions.set(clearEventId, INTERRUPTION_EVENT_KIND.AUDIO_CLEAR);
-      this.#send(cancelResponseEvents({ cancellationEventId, clearEventId }));
-    } else {
-      this.#pendingInterruptions.set(clearEventId, INTERRUPTION_EVENT_KIND.AUDIO_CLEAR);
-      this.#send(clearOutputAudioEvents(clearEventId));
-    }
-    // Then correct what Luke believes he said, or the next answer is free to
-    // refer back to a sentence that never reached the room.
-    if (truncateEvents.length > 0) {
-      this.#pendingInterruptions.set(truncationEventId, INTERRUPTION_EVENT_KIND.TRUNCATION);
-      this.#send(truncateEvents);
-    }
-    // The trim was this reply's last word: forgetting its item here is what
-    // stops the transcript still trailing in — the server had produced it
-    // before the cancel landed — from ever matching the caption again.
-    this.#responseItemId = undefined;
-    // And forgetting its response is what stops its finished form — cancelled
-    // or not, the server may already have completed it — from being read as
-    // the current turn's: a `response.done` that matches nothing can neither
-    // act with the new turn's arming nor end the new turn early.
-    this.#activeResponseId = undefined;
-    // The cancel concludes the reply at the server before anything sent after
-    // it is read — the channel is ordered — so nothing is outstanding from
-    // here, and whatever `done` the cancelled reply still sends matches no
-    // active response above.
-    this.#responseOutstanding = false;
-    this.#audioDrained = false;
-    this.#toolBatch = undefined;
-    this.#toolCallResponseIds.clear();
-    this.#generationDone = false;
-    this.#remoteQuiet = false;
-    this.#heardLuke = false;
-    this.#clearQuietTimer();
-    this.#clearSettleTimer();
-  }
-
-  /**
-   * Cuts off the reply under way without opening anything in its place — the
-   * developer asking for quiet rather than for a turn. The cut is the same
-   * one talking or typing over Luke makes: silenced at once, cancelled, and
-   * trimmed to what was actually heard, so the next answer cannot refer back
-   * to words that never reached the room. Reports whether there was a reply
-   * to stop, so the key that asked keeps its other meanings when there was
-   * not.
-   */
-  stopSpeaking(): boolean {
-    if (this.#status !== REALTIME_STATUS.RESPONDING) return false;
-    this.#interruptReply();
-    // A stop opens no reply of its own, so the turn moves on here: a tool
-    // follow-up still awaiting its write finds an epoch that is no longer its
-    // own and stands down, rather than speaking over the quiet just asked for.
-    this.#turnEpoch += 1;
-    this.#setStatus(REALTIME_STATUS.READY);
-    return true;
-  }
-
-  /**
-   * Voices one turn the main process decided — a briefing the brain handed
-   * the voice, or a scripted onboarding beat — reporting whether it could. A
-   * refusal is not a loss: the queue keeps the turn and tries again on its
-   * own clock, and a briefing that waits too long ages out rather than being
-   * read out as though it just happened.
-   */
-  speak(speech: ProactiveSpeechTurn): boolean {
-    if (speech.kind === ARRIVAL_SPEECH_KIND) {
-      const arrivalEvents = arrivalSpeechEvents(speech);
-      if (arrivalEvents.length === 0 || !this.isConnected || voiceExchangeActive(this.#status))
-        return false;
-      // No caption subject: the arrival speaks about no one observed session,
-      // like an introduction beat, so no notice may stand under the housing
-      // claiming it does.
-      this.#startResponse(arrivalEvents);
-      return true;
-    }
-    if (speech.kind === CALENDAR_ONBOARDING_SPEECH_KIND) {
-      if (!this.isConnected || voiceExchangeActive(this.#status)) return false;
-      // On the arrival's own terms: words with no kind, recorded as none.
-      this.#startResponse(calendarOnboardingSpeechEvents());
-      return true;
-    }
-    const events = briefingSpeechEvents(speech);
-    if (events.length === 0 || !this.isConnected || voiceExchangeActive(this.#status)) return false;
-    this.#startResponse(events);
-    // After the start, which clears the last reply's caption and kind: the
-    // briefing's reply is the one now under way until it ends.
-    this.#captionKind = REPLY_KIND.BRIEFING;
-    this.#emitCaption();
-    return true;
-  }
-
-  /**
-   * Voices one scripted beat of the introduction, reporting whether it could.
-   * The beat's direction is fixed by the build and its data already bounded;
-   * the turn opens with `tool_choice: "none"` on a session that declared no
-   * tools, so nothing about it can arm an act. No caption subject is set —
-   * the introduction speaks about no observed session.
-   */
-  speakIntroduction(line: IntroductionLine): boolean {
-    const events = introductionSpeechEvents(line);
-    if (events.length === 0 || !this.isConnected || voiceExchangeActive(this.#status)) return false;
-    this.#startResponse(events);
-    return true;
-  }
-
-  async close(): Promise<void> {
-    this.clearConversation();
-  }
-
-  /**
-   * Retires the call that owns the current conversation. Realtime items cannot
-   * be enumerated and deleted reliably, so Clear crosses a call boundary: the
-   * next developer turn receives a fresh server-side conversation as well as
-   * the caller's freshly emptied local history.
-   */
-  clearConversation(): void {
-    this.#closed = true;
-    this.#teardown();
-    this.#setStatus(REALTIME_STATUS.IDLE);
-  }
-
-  /**
-   * Releases the microphone and the peer connection without deciding what the
-   * session's status becomes. Every path that stops the call goes through here,
-   * so a failure can never leave the macOS microphone indicator lit with no way
-   * to turn it off.
-   */
-  #teardown(): void {
-    if (this.#tearingDown) return;
-    this.#tearingDown = true;
-    let teardownError: unknown;
-    const release = (action: () => void): void => {
-      try {
-        action();
-      } catch (error) {
-        teardownError ??= error;
-      }
-    };
-    const transport = this.#sdkTransport;
-    this.#sdkTransport = undefined;
-    release(() => transport?.close());
-    this.#remoteTrack = undefined;
-    // A fresh token before the capture retires: a device still opening for
-    // the attempt this teardown ends finds it stale and releases itself. The
-    // capture retires while the track is still known — so it can close it —
-    // and the channel is already down, so retirement hands no track to a
-    // sender: there is no call left to carry it, and the words the capture
-    // still held die with the attempt.
-    this.#attempt = Symbol("connect-attempt");
-    release(() => this.#retirePressCapture());
-    this.#microphone = undefined;
-    this.#microphoneSender = undefined;
-    this.#silenceTrack = undefined;
-    const stream = this.#stream;
-    this.#stream = undefined;
-    let tracks: readonly MediaStreamTrack[] = [];
-    release(() => {
-      tracks = stream?.getTracks() ?? [];
-    });
-    for (const track of tracks) release(() => track.stop());
-    // The reply's last words are handed over before the stores empty: the
-    // handover's write-back re-enters this session, and landing it here means
-    // the roster it renders against still stands — and everything it wrote is
-    // cleared with the rest below, so a retired call keeps nothing pending.
-    release(() => this.#clearCaption());
-    // What was said on the call goes with the call. The pending answers go too:
-    // they were built from stores this teardown is emptying, and the next call
-    // is filled from the app afresh before it takes a turn.
-    this.#pendingInterruptions.clear();
-    this.#responseOutstanding = false;
-    this.#audioDrained = false;
-    this.#toolBatch = undefined;
-    this.#toolCallResponseIds.clear();
-    this.#generationDone = false;
-    this.#remoteQuiet = false;
-    this.#heardLuke = false;
-    this.#clearQuietTimer();
-    this.#pendingTurn = false;
-    // The next call is minted at the stored pace, so nothing is owed to it.
-    this.#pendingSpeed = undefined;
-    this.#responseItemId = undefined;
-    this.#activeResponseId = undefined;
-    this.#audibleSince = undefined;
-    // Learned about this call, so it does not outlive it.
-    this.#audioEndingsReported = false;
-    this.#clearSettleTimer();
-    release(() => this.#options.onLocalStream(undefined));
-    release(() => this.#options.onRemoteStream(undefined));
-    this.#tearingDown = false;
-    if (teardownError !== undefined) {
-      try {
-        this.#options.onError(
-          teardownError instanceof Error ? teardownError.message : String(teardownError),
-        );
-      } catch {
-        // Teardown has already completed; an error reporter cannot undo it.
-      }
-    }
-  }
-
-  #startResponse(
-    events: readonly WireRecord[],
-    { keepCaption = false }: { keepCaption?: boolean } = {},
-  ): void {
-    // A pace still waiting from the last reply lands here, ahead of the
-    // request: the channel is ordered and no response is in progress — a
-    // cancel for the reply being talked over was sent before this — so the
-    // reply about to be asked for is already spoken at the new pace.
-    this.#flushPendingSpeed();
-    // The track is deliberately left as it is. A reply that was cut off left it
-    // disabled, and re-opening it here would let the tail of that reply — still
-    // arriving, because the server sent it before it was told to stop — be
-    // heard as the answer to what was just said. It is opened again when the
-    // server confirms the new reply has started, by which point the old one is
-    // certainly over: the data channel is ordered, so the clear that ended it
-    // was handled before the request for this one.
-    this.#generationDone = false;
-    this.#remoteQuiet = false;
-    this.#heardLuke = false;
-    // The reply now being asked for is the server's until it concludes it:
-    // nothing else may ask for one over it — and whatever follow-up was being
-    // waited on, this is the reply that answers or supersedes the wait.
-    this.#responseOutstanding = true;
-    this.#audioDrained = false;
-    this.#toolBatch = undefined;
-    this.#toolCallResponseIds.clear();
-    this.#clearQuietTimer();
-    this.#responseItemId = undefined;
-    // Nothing has been confirmed for this turn yet: whatever `response.done`
-    // arrives before the server confirms this reply belongs to a superseded
-    // one, and must find no active response to match.
-    this.#activeResponseId = undefined;
-    this.#audibleSince = undefined;
-    // A new turn: any brain follow-up still awaiting from the last turn will
-    // see this and stand down.
-    this.#turnEpoch += 1;
-    // A new turn starts with a clean strip; a follow-up continuing the same
-    // exchange keeps the words just said, and its own words stack under them.
-    if (!keepCaption) this.#clearCaption();
-    this.#clearSettleTimer();
-    if (events.some((event) => event.type === REALTIME_CLIENT_EVENT.INPUT_AUDIO_BUFFER_COMMIT)) {
-      this.#options.onSpokenAskClosed?.();
-    }
-    this.#send(events);
-    this.#setStatus(REALTIME_STATUS.RESPONDING);
-  }
-
-  /**
-   * The meter's report of whether Luke is audible. The meter calls quiet after
-   * a fifth of a second, which is shorter than the pause between two sentences,
-   * so a turn that ended on that edge would take the meter down mid-reply. The
-   * session waits for a silence longer than speech leaves behind, and ignores
-   * quiet that is not Luke's to answer for.
-   */
-  reportRemoteAudioLevel(active: boolean): void {
-    this.#clearQuietTimer();
-    if (this.#status !== REALTIME_STATUS.RESPONDING) return;
-    if (active) {
-      this.#heardLuke = true;
-      this.reportRemoteAudioActive();
-      return;
-    }
-    this.#quietTimer = setTimeout(() => {
-      this.#quietTimer = undefined;
-      // Only Luke's own silence ends Luke's turn.
-      if (!quietIsLukesOwn({ status: this.#status, heardLuke: this.#heardLuke })) return;
-      this.reportRemoteAudioIdle();
-    }, REMOTE_QUIET_MS);
-  }
-
-  /**
-   * Reports that Luke's audio has gone quiet, after the caller has already
-   * decided the silence is his and has lasted long enough to be an ending.
-   */
-  reportRemoteAudioIdle(): void {
-    // Remembered rather than acted on and forgotten: if generation has not
-    // finished yet, this is still the only quiet edge the meter will report,
-    // and `response.done` is what will read it.
-    this.#remoteQuiet = true;
-    // Nothing to infer on a call that reports its own endings. Inferring here
-    // is what ended a reply in the pause between its two sentences.
-    if (this.#audioEndingsReported) return;
-    if (!this.#generationDone) return;
-    this.#finishResponse();
-  }
-
-  /**
-   * Reports that Luke is audible again. A pause between two sentences is longer
-   * than the meter's idea of quiet, so without this a reply that pauses and
-   * resumes would end on the pause the moment generation finished — with Luke
-   * still speaking.
-   */
-  reportRemoteAudioActive(): void {
-    this.#remoteQuiet = false;
-    // The first time this reply is heard is the clock a truncate measures
-    // against. Later edges are pauses within it, not new beginnings.
-    this.#audibleSince ??= this.#now();
-  }
-
-  /**
-   * What to trim the cut-off reply to, if there is anything to trim. Nothing
-   * heard means nothing to correct — a reply interrupted in the gap before its
-   * first word left no impression to undo — and a reply whose audio already
-   * ran out was heard whole: the record needs no correction, and the wall
-   * clock has been counting past the audio's end for as long as the turn has
-   * held for its `done`, so a trim measured from it would ask past the end
-   * and be refused.
-   */
-  #truncateEvents(truncationEventId: string): readonly WireRecord[] {
-    const itemId = this.#responseItemId;
-    const audibleSince = this.#audibleSince;
-    if (!itemId || audibleSince === undefined || this.#currentReplyDrained()) return [];
-    return truncateResponseEvents({
-      itemId,
-      audioEndMs: this.#now() - audibleSince,
-      truncationEventId,
-    });
-  }
-
-  /**
-   * Whether the reply now under way has played out every word it generated.
-   * Only a drain that is attributably this reply's own says so: an old
-   * reply's late drain speaks for audio the current reply never played. A
-   * drain that named no reply keeps the old reading — it is nearly always
-   * the current one's, and reading it as another's would hold the turn to
-   * the settle backstop.
-   */
-  #currentReplyDrained(): boolean {
-    if (this.#audioDrained === false) return false;
-    const { responseId } = this.#audioDrained;
-    return responseId === undefined || responseId === this.#activeResponseId;
-  }
-
-  #now(): number {
-    return this.#options.now?.() ?? performance.now();
-  }
-
-  #silenceLuke(): void {
-    if (this.#remoteTrack) this.#remoteTrack.enabled = false;
-  }
-
-  #unsilenceLuke(): void {
-    if (this.#remoteTrack) this.#remoteTrack.enabled = true;
-  }
-
-  /** Starts the backstop for a reply whose proper ending never arrives. */
-  #armSettleTimer(delayMs: number = REALTIME_SETTLE_TIMEOUT_MS): void {
-    this.#settleTimer ??= setTimeout(() => {
-      this.#settleTimer = undefined;
-      this.#finishResponse();
-    }, delayMs);
-  }
-
-  #clearSettleTimer(): void {
-    if (this.#settleTimer === undefined) return;
-    clearTimeout(this.#settleTimer);
-    this.#settleTimer = undefined;
-  }
-
-  #clearQuietTimer(): void {
-    if (this.#quietTimer === undefined) return;
-    clearTimeout(this.#quietTimer);
-    this.#quietTimer = undefined;
-  }
-
-  #clearCaption(): void {
-    // The kind is of the reply, so the reply ending takes it too: every path
-    // that ends one clears the caption through here.
-    if (this.#captionSegments.length === 0 && this.#captionKind === undefined) return;
-    // Every path that ends a reply passes here, so this is where its words
-    // are handed over before they are let go: the one moment they are both
-    // final and still known.
-    const texts = this.#captionTexts();
-    const kind = this.#captionKind;
-    const runId = this.#captionRunId;
-    this.#captionSegments = [];
-    this.#captionKind = undefined;
-    this.#captionRunId = undefined;
-    // A reply that said nothing leaves no words to hand over — unless it was
-    // voicing a run's end, whose ending is owed to the delivery it was granted
-    // under however little of it was heard.
-    if (texts || runId !== undefined) this.#options.onReplyEnded?.(texts ?? [], kind, runId);
-    this.#options.onCaption(undefined, undefined);
-  }
-
-  /** Clears an undelivered briefing without admitting it to History. */
-  #discardCaption(): void {
-    this.#captionSegments = [];
-    this.#captionKind = undefined;
-    this.#captionRunId = undefined;
-    this.#options.onCaption(undefined, undefined);
-  }
-
-  /** What the caption currently says, or undefined with nothing to say. */
-  #captionTexts(): readonly string[] | undefined {
-    if (this.#captionSegments.length === 0) return undefined;
-    return this.#captionSegments.map((segment) => segment.text);
-  }
-
-  #emitCaption(): void {
-    this.#options.onCaption(this.#captionTexts(), this.#captionKind);
-  }
-
-  /**
-   * Grows the caption with the words just generated. The current item's words
-   * grow its own segment; an item taking over from another — the reply's
-   * second message, or the follow-up after a tool call — starts a segment of
-   * its own, so two responses stack instead of running together. Only the
-   * newest {@link CAPTION_SEGMENT_LIMIT} stay up.
-   */
-  #appendCaptionDelta(itemId: string | undefined, delta: string): void {
-    const last = this.#captionSegments.at(-1);
-    if (last && last.itemId === itemId) {
-      last.text += delta;
-    } else {
-      this.#captionSegments.push({ itemId, text: delta });
-      this.#captionSegments = this.#captionSegments.slice(-CAPTION_SEGMENT_LIMIT);
-    }
-    this.#emitCaption();
-  }
-
-  /**
-   * Lands an item's final transcript on the segment its deltas built — even
-   * after a later item has taken the turn on, which is what keeps a settled
-   * response's words whole while the next one streams under them. A transcript
-   * whose item holds no segment is a cancelled reply's straggler, and writes
-   * nothing.
-   */
-  #settleCaptionTranscript(itemId: string | undefined, transcript: string): void {
-    const segment = this.#captionSegments.find((candidate) => candidate.itemId === itemId);
-    if (!segment || segment.text === transcript) return;
-    segment.text = transcript;
-    this.#emitCaption();
-  }
-
-  /** Ends the turn once the reply is done, so the next one can start. */
-  #finishResponse(): void {
-    this.#generationDone = false;
-    // However the turn ended — the settle backstop included — whatever the
-    // server still owed it is treated as concluded, so a `done` that never
-    // comes cannot leave every later reply refused against it.
-    this.#responseOutstanding = false;
-    this.#audioDrained = false;
-    this.#toolBatch = undefined;
-    this.#toolCallResponseIds.clear();
-    // The turn is over, and everything of it is spent — a write still in
-    // flight from it finds this boundary and stands down, rather than opening
-    // its follow-up out of a silence already declared.
-    this.#turnEpoch += 1;
-    // No reply is current once the turn is over: a `done` that outlives the
-    // settle backstop reads as a stranger's, and nothing of it reaches the
-    // turn the developer was already told had ended.
-    this.#activeResponseId = undefined;
-    // The caption is of speech, and the speech is over. Whatever ended the
-    // reply — the audio draining, an error, the settle timer — the words leave
-    // with the meter and the face rather than lingering under a quiet capsule.
-    this.#clearCaption();
-    // Whatever ended the reply — an error, the settle timer, Luke simply
-    // stopping — the next one has to be audible. Without this a reply that
-    // failed before it started would leave Luke silenced with nothing to
-    // un-silence him.
-    this.#unsilenceLuke();
-    this.#clearQuietTimer();
-    this.#clearSettleTimer();
-    this.#heardLuke = false;
-    // The reply is over, so the API is between turns — the one moment it
-    // accepts a pace change that arrived while Luke was speaking.
-    this.#flushPendingSpeed();
-    if (this.#status === REALTIME_STATUS.RESPONDING) this.#setStatus(REALTIME_STATUS.READY);
-    this.#options.onReplySettled?.();
-  }
-
-  /**
-   * Changes how fast Luke speaks on the call now open, from his next reply on.
-   * A call minted at one pace stays a live session, so the change travels as a
-   * session update rather than waiting for the next conversation. The API
-   * applies a pace only between model turns: a change landing mid-reply is
-   * held and sent when the reply ends. With no call open there is nothing to
-   * update — the next one is minted at the stored pace already.
-   */
-  applySpeed(speed: number): void {
-    if (!this.isConnected) {
-      // A call being opened was minted at whatever pace stood when its
-      // credential was asked for, which this change may already have
-      // overtaken: hold it and send it once the channel opens. Sent to a
-      // call that was minted at the new pace after all, it is a no-op.
-      if (this.isConnecting) this.#pendingSpeed = speed;
-      return;
-    }
-    if (this.#status === REALTIME_STATUS.RESPONDING) {
-      this.#pendingSpeed = speed;
-      return;
-    }
-    this.#pendingSpeed = undefined;
-    this.#send(outputSpeedUpdateEvents(speed));
-  }
-
-  /** Sends the pace change that waited out a reply, once nothing is speaking. */
-  #flushPendingSpeed(): void {
-    const speed = this.#pendingSpeed;
-    if (speed === undefined) return;
-    this.#pendingSpeed = undefined;
-    this.#send(outputSpeedUpdateEvents(speed));
-  }
-
-  /**
-   * Handles an error answering one interruption without letting an old reply's
-   * failure finish the new turn that interrupted it. The documented
-   * no-active-response race is quiet; every other refusal still reaches the
-   * developer as a real voice error.
-   */
-  #interruptionError(event: { message: string; eventId?: string; errorType?: string }): boolean {
-    if (event.eventId === undefined) return false;
-    const kind = this.#pendingInterruptions.get(event.eventId);
-    if (kind === undefined) return false;
-    this.#pendingInterruptions.delete(event.eventId);
-    const benignCancellation =
-      kind === INTERRUPTION_EVENT_KIND.CANCELLATION &&
-      event.errorType === "invalid_request_error" &&
-      NO_ACTIVE_RESPONSE_CANCELLATION.test(event.message);
-    if (!benignCancellation) this.#options.onError(event.message);
-    return true;
-  }
-
-  get #toolFollowUpPending(): boolean {
-    const batch = this.#toolBatch;
-    return Boolean(batch?.responseDone && !batch.followUpStarted && batch.callIds.size > 0);
-  }
-
-  #observeSdkToolCall(record: WireRecord): void {
-    if (record.type !== "response.output_item.done") return;
-    const item = isRecord(record.item) ? record.item : undefined;
-    if (item?.type !== "function_call") return;
-    const responseId = text(record.response_id);
-    const callId = text(item.call_id);
-    if (!responseId || !callId) return;
-    this.#toolCallResponseIds.set(callId, responseId);
-    const batch = this.#toolBatch;
-    if (batch?.responseId === responseId) batch.callIds.add(callId);
-  }
-
-  #handleServerEvent(data: UnparsedWireValue): void {
-    // Decoded once, here: the tap reads the record whole — the parser below
-    // keeps only the events the conversation acts on, and what it discards
-    // (usage, errors in full, event types this build does not know) is half
-    // of what a trace exists to show — and the parser accepts the decoded
-    // record as readily as the string, so nothing is parsed twice.
-    const record = decodeRealtimePayload(data);
-    if (record) {
-      this.#options.onWireEvent?.(TRACE_DIRECTION.SERVER, record);
-      this.#observeSdkToolCall(record);
-    }
-    const event = parseRealtimeServerEvent(record);
-    if (!event) return;
-
-    switch (event.type) {
-      case REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_ITEM_ADDED:
-        if (event.itemId) this.#responseItemId = event.itemId;
-        return;
-      case REALTIME_SERVER_EVENT.RESPONSE_CREATED:
-        // Only a reply the session is still waiting on may be adopted. A
-        // confirmation arriving after the developer stopped or took the turn is
-        // the cancelled reply's own, landing late — the stop raced the server's
-        // confirmation — and adopting it would re-open the track over the quiet
-        // just asked for, and let its finished form read as the current reply's.
-        if (this.#status !== REALTIME_STATUS.RESPONDING) return;
-        // The reply being asked for is under way, so anything arriving from here
-        // belongs to it rather than to the one it replaced. Its name is what a
-        // `response.done` must present to be read as this reply's: the channel
-        // is ordered, so a cancelled reply's `done` lands before this
-        // confirmation and finds nothing to match.
-        if (event.responseId) {
-          this.#activeResponseId = event.responseId;
-          this.#toolBatch = {
-            responseId: event.responseId,
-            callIds: new Set(),
-            outputCallIds: new Set(),
-            epoch: this.#turnEpoch,
-            responseDone: false,
-            followUpStarted: false,
-          };
-        }
-        this.#unsilenceLuke();
-        return;
-      case REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_AUDIO_TRANSCRIPT_DELTA:
-        // Only the reply being spoken may write the caption. A cancelled reply's
-        // transcript keeps arriving after the interrupt that cleared it — the
-        // server had already produced it — and without this check a late piece
-        // would draw the words Luke was just stopped from saying, or splice them
-        // onto the next reply's.
-        if (event.itemId === this.#responseItemId && event.delta) {
-          this.#appendCaptionDelta(event.itemId, event.delta);
-        }
-        return;
-      case REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_AUDIO_TRANSCRIPT_DONE:
-        // The server's own rendering of the whole item, which the deltas only
-        // approximate: a delta lost to the channel would otherwise leave a hole
-        // in the sentence for as long as it stayed up. It lands on the segment
-        // the item's deltas built — even one the turn has already moved past —
-        // and a cancelled reply's `done`, the likeliest straggler of all, finds
-        // its segments cleared and writes nothing.
-        if (event.transcript) {
-          this.#settleCaptionTranscript(event.itemId, event.transcript);
-        }
-        return;
-      case REALTIME_SERVER_EVENT.OUTPUT_AUDIO_BUFFER_STARTED:
-        // Audio flowing again says the drain the turn remembered was the
-        // pause between two things the reply had to say, not its ending.
-        // Left standing, the stale drain lets the `done` end the turn under
-        // the second half — the face and the duck released while Luke is
-        // still speaking. Only a resume that is attributably the current
-        // reply's own — or unnamed, the same reading the drain gets — may
-        // un-remember it. The backstop a drain armed is restarted rather
-        // than kept or cleared: kept, its clock ran from the pause and cuts
-        // a resumed half that outlives it; cleared, a `done` that never
-        // comes would hold the turn open with nothing left to end it.
-        if (event.responseId !== undefined && event.responseId !== this.#activeResponseId) {
-          return;
-        }
-        this.#audioDrained = false;
-        if (this.#settleTimer !== undefined) {
-          this.#clearSettleTimer();
-          this.#armSettleTimer();
-        }
-        return;
-      case REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_DELTA:
-        // The completed transcript's own microphone guard, applied to its
-        // preview: a speak-only call has no developer words to be taking
-        // shape either.
-        if (this.#withMicrophone) this.#options.onSpokenAskDelta?.(event.itemId, event.delta);
-        return;
-      case REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_COMPLETED:
-        // Only the developer's own call has spoken turns to hand back; the
-        // guard is belt to the speak-only shape's suspenders, so a stray
-        // event on Luke's own call can never write a developer line.
-        if (this.#withMicrophone) this.#options.onSpokenAsk?.(event.transcript, event.itemId);
-        return;
-      case REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_FAILED:
-        if (this.#withMicrophone) this.#options.onSpokenAskFailed?.(event.itemId);
-        return;
-      case REALTIME_SERVER_EVENT.INPUT_AUDIO_BUFFER_COMMITTED:
-        if (this.#withMicrophone) this.#options.onSpokenAskCommitted?.(event.itemId);
-        return;
-      case REALTIME_SERVER_EVENT.OUTPUT_AUDIO_BUFFER_STOPPED:
-        this.#audioEndingsReported = true;
-        // The audio can run out while the server still owes the reply its
-        // `done` — generation finishing and playback finishing have no fixed
-        // order — and until that `done` the conversation holds an active
-        // response. A turn ended here would offer READY to a caller with a
-        // reply to ask for — the announcer reading out a notice that queued
-        // behind this reply is the one that takes it — and the create it
-        // sends would be refused as a conversation already in progress, the
-        // refusal read out as a voice error and the notice lost. So the
-        // drain is remembered and the `done` ends the turn, with the settle
-        // backstop for a `done` that never comes.
-        // An armed reply whose `done` has landed but whose follow-up is still
-        // owed holds the same way, in the mirror order: the write is under
-        // way, the READY an ending would offer is the same edge, and the
-        // `done` already gave the hold a clock of its own.
-        if (this.#responseOutstanding || this.#toolFollowUpPending) {
-          this.#audioDrained = { responseId: event.responseId };
-          this.#armSettleTimer();
-          return;
-        }
-        // The reply is over because the server says the audio ran out, not
-        // because this end guessed from a stretch of quiet. A pause between two
-        // sentences is quiet too, and guessing ended the turn in the middle of
-        // one — the meter and the face went with it while Luke talked on.
-        this.#finishResponse();
-        return;
-      case REALTIME_SERVER_EVENT.RESPONSE_DONE: {
-        // Whether this is the reply now under way, or the finished form of one
-        // the developer already talked or typed over. The server had completed
-        // the old reply before the cancel landed — it generates ahead of the
-        // room — so its `done` still arrives, after the interrupt has already
-        // opened a new turn. Nothing of it may reach the brain as that turn's
-        // ask or end that turn early: its calls are answered refused so the
-        // model is not left waiting, and everything else about it is ignored.
-        const fresh = event.responseId === this.#activeResponseId;
-        // Whatever this reply turns out to be below, the server has concluded
-        // it: from here the conversation can take a new `response.create`.
-        if (fresh) this.#responseOutstanding = false;
-        // A reply that asked for tools has not finished talking: the calls are
-        // answered and the reply resumes over their outcomes, so the turn stays
-        // open rather than ending on a reply that was only half made.
-        if (event.calls.length > 0) {
-          let batch: SdkToolBatch | undefined;
-          if (fresh && event.responseId) {
-            batch =
-              this.#toolBatch?.responseId === event.responseId
-                ? this.#toolBatch
-                : {
-                    responseId: event.responseId,
-                    callIds: new Set<string>(),
-                    outputCallIds: new Set<string>(),
-                    epoch: this.#turnEpoch,
-                    responseDone: false,
-                    followUpStarted: false,
-                  };
-          }
-          if (batch && batch.responseId === event.responseId) {
-            this.#toolBatch = batch;
-            batch.responseDone = true;
-            for (const call of event.calls) {
-              batch.callIds.add(call.callId);
-              this.#toolCallResponseIds.set(call.callId, batch.responseId);
-            }
-          }
-          if (batch) {
-            // The turn now holds for the follow-up, because the READY an
-            // ending here would offer while the brain thinks is the edge the
-            // queue rides — a briefing taken there bumps the epoch, and the
-            // follow-up voicing the answer stands down against it, the
-            // developer's answer abandoned for a briefing. The hold is the
-            // ask's, so it gets a clock of its own, long enough for a brain
-            // turn that reads and acts before it answers — while an ask that
-            // hangs past even that still meets a backstop, because a turn
-            // that never ends is worse than one that ends early.
-            this.#clearSettleTimer();
-            this.#armSettleTimer(BRAIN_ASK_SETTLE_TIMEOUT_MS);
-            this.#startToolFollowUpIfReady();
-            return;
-          }
-          // The spoken half's audio already drained — its ending deferred to
-          // this `done`, and a reply owing no follow-up ends here, exactly
-          // as the drain would have ended it.
-          if (fresh && this.#currentReplyDrained()) this.#finishResponse();
-          return;
-        }
-        if (!fresh) return;
-        // A reply the server says made no sound has nothing to play out — a
-        // success is said with silence, so the follow-up after a tool call is
-        // often exactly this. The meter will never hear him and never call
-        // him quiet, and waiting out the settle backstop would hold the meter
-        // and the face on a reply that was over the moment it was finished.
-        // Only a response that reported its output may end here: an unknown
-        // is not a silence, and keeps the ordinary endings below.
-        if (event.hasAudio === false) {
-          this.#finishResponse();
-          return;
-        }
-        // Generation is done; the reply is not. The turn ends when Luke stops
-        // being audible, which the caller reports from the audio itself rather
-        // than from an event — the one that would say so is undocumented.
-        this.#generationDone = true;
-        // The server said the audio ran out before it said the reply was
-        // over. That ending waited for this `done` — the conversation held
-        // an active response until it — and lands now.
-        if (this.#currentReplyDrained()) {
-          this.#finishResponse();
-          return;
-        }
-        // The audio can run out before the event that says generation is over.
-        // The meter has already reported its quiet and will not report it twice,
-        // so waiting for another would hold the turn open until the settle
-        // timeout — seconds of a meter and a face saying Luke is still talking.
-        if (this.#remoteQuiet && !this.#audioEndingsReported) {
-          this.#finishResponse();
-          return;
-        }
-        this.#armSettleTimer();
-        return;
-      }
-      case REALTIME_SERVER_EVENT.ERROR:
-        // Only Luke's own trim can draw a past-the-end refusal, the service
-        // clamps and truncates anyway, and it names no event this could match
-        // it by — `error.event_id` is null on the wire. Recognized by its
-        // sentence and never shown.
-        if (TRUNCATION_PAST_AUDIO_END.test(event.message)) return;
-        if (this.#interruptionError(event)) return;
-        this.#options.onError(event.message);
-        // An error can arrive *instead of* `response.done` — an empty push-to-talk
-        // commit is the common case — which would otherwise leave the session
-        // stuck in `responding` and unable to take another turn. But only a
-        // reply the server never confirmed ends this way: behind a confirmed
-        // one an error is an aside — the reply is still the server's, its own
-        // `done` still ends the turn, and ending it here would offer READY
-        // while the conversation still holds an active response. The settle
-        // backstop covers a `done` that never comes.
-        if (this.#activeResponseId !== undefined) {
-          this.#armSettleTimer();
-          return;
-        }
-        this.#finishResponse();
-    }
-  }
-
-  async #sdkToolOutput(
-    expectedName: string,
-    details: SdkToolCallDetails | undefined,
-  ): Promise<WireRecord> {
-    const toolCall = details?.toolCall;
-    const callId = toolCall?.callId;
-    const name = toolCall?.name;
-    if (toolCall?.type !== "function_call" || !callId || name !== expectedName) {
-      return { status: ACT_RESULT_STATUS.REJECTED, reason: "The tool call was malformed." };
-    }
-    const argumentsJson = toolCall.arguments;
-    if (argumentsJson === undefined) {
-      return { status: ACT_RESULT_STATUS.REJECTED, reason: "The tool arguments were malformed." };
-    }
-    const responseId = this.#toolCallResponseIds.get(callId);
-    const batch = responseId === this.#toolBatch?.responseId ? this.#toolBatch : undefined;
-    // Only the reply now under way may ask the brain: a cancelled reply's
-    // late call — the developer already talked over it — is answered with a
-    // refusal rather than an ask the developer moved on from.
-    const current = Boolean(batch && batch.epoch === this.#turnEpoch && batch.callIds.has(callId));
-    return this.#toolCallOutput({ name, callId, argumentsJson }, current);
-  }
-
-  #toolOutputSent(callId: string): void {
-    const responseId = this.#toolCallResponseIds.get(callId);
-    const batch = responseId === this.#toolBatch?.responseId ? this.#toolBatch : undefined;
-    if (!batch?.callIds.has(callId)) return;
-    batch.outputCallIds.add(callId);
-    this.#startToolFollowUpIfReady();
-  }
-
-  #startToolFollowUpIfReady(): void {
-    const batch = this.#toolBatch;
-    if (
-      !batch?.responseDone ||
-      batch.followUpStarted ||
-      batch.epoch !== this.#turnEpoch ||
-      !this.isConnected ||
-      [...batch.callIds].some((callId) => !batch.outputCallIds.has(callId))
-    ) {
-      return;
-    }
-    batch.followUpStarted = true;
-    this.#startResponse(functionCallFollowUpEvents(), { keepCaption: true });
-  }
-
-  /**
-   * Answers the voice's one tool. The developer's words go to the brain over
-   * the bridge, and what comes back — the reply, or a bounded refusal — is
-   * the tool's output, for the follow-up to say. The brain acts on its own
-   * side behind its own validators; nothing here performs anything.
-   */
-  async #toolCallOutput(call: ParsedRealtimeFunctionCall, current: boolean): Promise<WireRecord> {
-    if (!current) {
-      return {
-        status: ACT_RESULT_STATUS.REJECTED,
-        reason: "That turn is over; ask again if it still matters.",
-      };
-    }
-    if (call.name !== ASK_BRAIN_TOOL.name) {
-      return { status: ACT_RESULT_STATUS.REJECTED, reason: "No such tool exists." };
-    }
-    const question = askQuestion(call.argumentsJson);
-    if (!question) {
-      return { status: ACT_RESULT_STATUS.REJECTED, reason: "The ask carried no words." };
-    }
-    if (!this.#options.askBrain) {
-      return {
-        status: ACT_RESULT_STATUS.REJECTED,
-        reason: "Luke's judgment is not available on this call.",
-      };
-    }
-    // The ask can take a while — a brain turn reads and may act — so the
-    // turn's backstop is stretched to the ask's own clock for its duration.
-    this.#clearSettleTimer();
-    this.#armSettleTimer(BRAIN_ASK_SETTLE_TIMEOUT_MS);
-    const epoch = this.#turnEpoch;
-    let answer: BrainAskResult;
-    try {
-      answer = await this.#options.askBrain(question, call.callId);
-    } catch {
-      return {
-        status: ACT_RESULT_STATUS.REJECTED,
-        reason: "Luke's judgment did not answer.",
-      };
-    }
-    if (answer.status === BRAIN_ASK_PENDING_STATUS) {
-      // The run goes on without this turn: the follow-up says so, and the
-      // turn is released rather than held open for a reply that may be a
-      // long time coming.
-      return { status: answer.status, note: answer.note };
-    }
-    if (answer.status !== ACT_RESULT_STATUS.ACCEPTED) {
-      return { status: answer.status, reason: answer.reason };
-    }
-    // The follow-up that says the answer is the reply now under way, and its
-    // words are the brain's — marked here, before the follow-up opens with the
-    // caption kept. Unless the developer stopped or took the turn while the
-    // ask was out: the follow-up stands down on that edge, and a late answer
-    // must not mark words that will never be said. The answer itself still
-    // travels, because the brain did what it says it did.
-    if (epoch === this.#turnEpoch) {
-      this.#captionKind = REPLY_KIND.REPLY;
-      this.#captionRunId = answer.runId;
-      this.#emitCaption();
-    }
-    return { briefing: answer.briefing };
-  }
-
-  #send(events: readonly WireRecord[]): void {
-    const transport = this.#sdkTransport;
-    if (transport?.status !== "connected") return;
-    for (const event of events) transport.sendEvent(event);
-  }
-
-  #fail(message: string): boolean {
-    // Release the device before reporting. `FAILED` offers "Start voice" again,
-    // and retrying must not stack a second call on top of a live microphone.
-    if (this.#captionKind !== undefined) this.#discardCaption();
-    this.#teardown();
-    this.#options.onError(message);
-    this.#setStatus(REALTIME_STATUS.FAILED);
-    return false;
-  }
-
-  #setStatus(status: RealtimeStatus): void {
-    if (this.#status === status) return;
-    this.#status = status;
-    // The exchange settling is what closes the device the press opened — not
-    // the commit itself, because closing a capture device is audible on
-    // shared hardware (a Bluetooth headset renegotiates its codec), and at
-    // the commit Luke is just starting to answer. Here the reply is over and
-    // the blip lands in the quiet. A press already waiting keeps the device:
-    // its turn is about to reuse it.
-    if (status === REALTIME_STATUS.READY && !this.#pendingTurn) this.#releaseMicrophone();
-    this.#options.onStatus(status);
-  }
-}
